fix(terminal) change defaults for the instance terminal, raise state, so it is persisted when reconnecting multiple times
