+++ src/pages/instances/actions/ReconnectTerminalBtn.tsx
@@ -2,6 +2,8 @@ import React, { FC, useState } from "react";
 import { Button, Icon } from "@canonical/react-components";
 import { LxdTerminalPayload } from "types/terminal";
 import TerminalPayloadForm from "../TerminalPayloadForm";
+import * as Yup from "yup";
+import { useFormik } from "formik";
 
 interface Props {
   onFinish: (data: LxdTerminalPayload) => void;
@@ -10,8 +12,7 @@ interface Props {
 const ReconnectTerminalBtn: FC<Props> = ({ onFinish }) => {
   const [isModal, setModal] = useState(false);
 
-  const confirmModal = (result: LxdTerminalPayload) => {
-    onFinish(result);
+  const closeModal = () => {
     setModal(false);
   };
 
@@ -19,14 +20,56 @@ const ReconnectTerminalBtn: FC<Props> = ({ onFinish }) => {
     setModal(true);
   };
 
+  const TerminalSchema = Yup.object().shape({
+    command: Yup.string().required("This field is required"),
+    environment: Yup.array().of(
+      Yup.object().shape({
+        key: Yup.string(),
+        value: Yup.string(),
+      })
+    ),
+    user: Yup.number(),
+    group: Yup.number(),
+  });
+
+  const formik = useFormik({
+    initialValues: {
+      command: "bash",
+      environment: [
+        {
+          key: "TERM",
+          value: "xterm-256color",
+        },
+        {
+          key: "HOME",
+          value: "/root",
+        },
+      ],
+      user: 0,
+      group: 0,
+    },
+    validationSchema: TerminalSchema,
+    onSubmit: (values) => {
+      const result = {
+        command: [values.command],
+        "record-output": true,
+        "wait-for-websocket": true,
+        environment: values.environment.reduce(
+          (a, v) => ({ ...a, [v.key]: v.value }),
+          {}
+        ),
+        interactive: true,
+        group: values.group,
+        user: values.user,
+      };
+      onFinish(result);
+      closeModal();
+    },
+  });
+
   return (
     <>
-      {isModal && (
-        <TerminalPayloadForm
-          close={() => setModal(false)}
-          onFinish={confirmModal}
-        />
-      )}
+      {isModal && <TerminalPayloadForm close={closeModal} formik={formik} />}
       <Button className="u-no-margin--bottom" hasIcon onClick={openModal}>
         <Icon name="connected" />
         <span>Reconnect</span>
